+++ src/environment/room_map.ts
@@ -2,7 +2,7 @@ import { Warp } from "../warp";
 import { IRawMap } from "./maps";
 import { Tile, TILE_TYPES, TILE_REF } from "./tile";
 import { IDrawable } from "../idrawable";
-import { RoomWalls } from "./room_walls";
+import { RoomWalls } from "./walls/room_walls";
 
 export abstract class RoomMap implements IDrawable {
