+++ src/environment/rooms/four_fire_room.ts
@@ -1,7 +1,14 @@
 import { RoomMap } from "../room_map";
-import { FloorOneWalls } from "../wall_types/floor_one_room";
 import { IRawMap } from "../maps";
 import { Direction } from "../../enum";
+import { RoomWalls } from "../walls/room_walls";
+import { CustomWallElement } from "../walls/wall_element";
+import { RoomCornerWall } from "../walls/corner_wall";
+import { WallSprite } from "../walls/wall_sprite";
+import { Point } from "../../point";
+import { SideWall } from "../walls/side_wall";
+import { Door } from "../walls/door";
+import { canvas_H } from "../../main";
 
 export class FourFireRoom extends RoomMap {
 
@@ -30,7 +37,34 @@ export class FourFireRoom extends RoomMap {
 				]
 		};
 
-		super(raw_map, new FloorOneWalls(door_placement));
+		const corner_sprite = new WallSprite(new Point(0, 0), new Point(60, 60), "assets/img/walls/floor_one.png");
+		const side_sprite = new WallSprite(new Point(60, 0), new Point(120, 60), "assets/img/walls/floor_one.png");
+		const open_door_sprite = new WallSprite(new Point(120, 0), new Point(180, 60), "assets/img/walls/floor_one.png");
+		const closed_door_sprite = new WallSprite(new Point(240, 0), new Point(300, 60), "assets/img/walls/floor_one.png");
+		const cracks = new WallSprite(new Point(180, 0), new Point(240, 60), "assets/img/walls/floor_one.png");
+
+		const side_walls = [
+			new SideWall(Direction.UP, side_sprite),
+			new SideWall(Direction.DOWN, side_sprite),
+			new SideWall(Direction.LEFT, side_sprite),
+			new SideWall(Direction.RIGHT, side_sprite),
+		];
+
+		const corner_walls = [
+			new RoomCornerWall(Direction.TOP_LEFT, corner_sprite),
+			new RoomCornerWall(Direction.TOP_RIGHT, corner_sprite),
+			new RoomCornerWall(Direction.BOTTOM_LEFT, corner_sprite),
+			new RoomCornerWall(Direction.BOTTOM_RIGHT, corner_sprite),
+		];
+
+		const doors = door_placement.map((placement, index) => new Door(placement, open_door_sprite, closed_door_sprite, index % 2 === 0));
+
+		const custom_elements = [
+			new CustomWallElement(Direction.UP, cracks, new Point(180, 0)),
+			new CustomWallElement(Direction.DOWN, cracks, new Point(360, canvas_H - cracks.height))
+		];
+
+		super(raw_map, new RoomWalls(side_walls, corner_walls, doors, custom_elements));
 	}
 
 }
