+++ src/environment/walls/wall_sprite.ts
@@ -1,4 +1,4 @@
-import { Point } from "../point";
+import { Point } from "../../point";
 
 export class WallSprite {
