+++ src/main.ts
@@ -3,7 +3,7 @@ import { Renderer } from "./renderer";
 import { GameState } from "./gamestate";
 import { Settings } from "./settings/settings";
 import { KeyMapper } from "./settings/keymapper";
-import { FourFireRoom } from "./environment/map_types/four_fire_room";
+import { FourFireRoom } from "./environment/rooms/four_fire_room";
 import { Direction } from "./enum";
 
 export const canvas = document.querySelector("#canvas") as HTMLCanvasElement;
@@ -13,7 +13,7 @@ export const ctx = canvas.getContext("2d");
 export const bank = new ImageBank();
 export const renderer = new Renderer();
 export const gameState = new GameState(
-	new FourFireRoom([Direction.UP, Direction.LEFT, Direction.RIGHT, Direction.DOWN])
+	new FourFireRoom([Direction.UP, Direction.LEFT, Direction.RIGHT])
 );
 
 //TODO: add a localstorage service to retrieve the value the user wants to use
